Hardware signing WIP #3
